com.safetynet.safetynetalertsapi.exceptions.InvalidAddressException.java: move to proper package.
com.safetynet.safetynetalertsapi.exceptions.IdentityMissmatchException: add custom exception (firstname and lastname in pathvariables dont't match with Identity properties of an object)
