+++ src/main/java/com/safetynet/safetynetalertsapi/exceptions/IdentityMismatchException.java
@@ -1,0 +1,7 @@
+package com.safetynet.safetynetalertsapi.exceptions;
+
+public class IdentityMismatchException extends Throwable {
+    public IdentityMismatchException(String message) {
+        super(message);
+    }
+}
