+++ src/main/java/com/safetynet/safetynetalertsapi/exceptions/InvalidAddressException.java
@@ -1,4 +1,4 @@
-package com.safetynet.safetynetalertsapi.repositories;
+package com.safetynet.safetynetalertsapi.exceptions;
 
 public class InvalidAddressException extends Throwable {
     public InvalidAddressException(String message) {
